refactor(web): split the user-tool composers from their chat bindings
CalendarChatCompose and the email ComposeTool mixed the composer (the event
form, the email compose layout) with how chat finishes a deferred call
(cognition tool endpoints, setMessages). The composers now take a
UserToolReviewSink - canAct, lockedNotice, onEdit, onExecute, onReject - and
the chat wrappers build the sink over the same endpoints as before, so an
agent session can mount the same composers over a review elicitation.
